// packages/extensions/src/extensions/ImageEditor/editor/types.ts
export interface CropRect {
  x: number;
  y: number;
  width: number;
  height: number;
}